+++ server-v2/models/employee.js
@@ -1,5 +1,7 @@
-const moment = require('moment');
-const uuid = require('uuid/v1');
+import moment from 'moment';
+import uuid from 'uuid/v1';
+import db from '../config/connection';
+
 
 class Employee {
   constructor(firstname, lastname, email, username, password, address, gender, jobrole,
@@ -16,6 +18,50 @@ class Employee {
     this.department = department;
     this.isadmin = false;
     this.createdon = moment().format('YYYY-MM-DD');
+  }
+
+  static async createUserTable() {
+    const result = await db.query(`
+          CREATE TABLE IF NOT EXISTS employee (
+              id UUID PRIMARY KEY,
+              firstname text,
+              lastname text,
+              email text UNIQUE,
+              address text,
+              gender text,
+              jobrole text,
+              department text,
+              isadmin boolean,
+              username text,
+              password text,
+              createdon date
+          )`);
+    return result;
+  }
+
+  static async createUser(data) {
+    const result = await db.query(`INSERT INTO employee(id, firstname, lastname, email, username, password, address, gender, jobrole,
+        department, isadmin, createdon) VALUES('${data.id}',
+        '${data.firstname}',
+        '${data.lastname}',
+        '${data.email}',
+        '${data.username}',
+        '${data.password}',
+        '${data.address}',
+        '${data.gender}',
+        '${data.jobrole}',
+        '${data.department}',
+        '${data.isadmin}',
+        '${data.createdon}'
+        ) returning *;
+      `);
+
+    return result.error;
+  }
+
+  static async CleanEmployee() {
+    const result = await db.query('DELETE FROM employee;');
+    return result;
   }
 }
